+++ core/src/builtin_widgets/svg.rs
@@ -2,9 +2,8 @@ use crate::prelude::*;
 
 impl Render for Svg {
   #[inline]
-  fn perform_layout(&self, _: BoxClamp, _: &mut LayoutCtx) -> Size { self.size }
+  fn perform_layout(&self, clamp: BoxClamp, _: &mut LayoutCtx) -> Size { clamp.clamp(self.size) }
 
-  #[inline]
   fn paint(&self, ctx: &mut PaintingCtx) {
     let painter = ctx.painter();
     painter.draw_svg(self);