+++ macros/src/declare_derive.rs
@@ -483,7 +483,10 @@ pub(crate) fn declare_derive(input: &mut syn::DeriveInput) -> syn::Result<TokenS
           self
         }
 
-        #[doc="Initializes the relative anchor to the parent of the widget"]
+        #[doc="Anchor child constraints relative to the parent widget. It's \
+        important to note that if you anchor the child widget outside of its \
+        parent, it may become unable to click, so ensure there is ample space \
+        within the parent."]
         #vis fn anchor<const _M: u8>(mut self, v: impl DeclareInto<Anchor, _M>) -> Self {
           self.fat_obj = self.fat_obj.anchor(v);
           self