+++ widgets/src/layout/only_sized_by_parent.rs
@@ -6,6 +6,12 @@ use ribir_core::prelude::*;
 #[derive(SingleChild, Declare)]
 pub struct OnlySizedByParent {}
 
+// `OnlySizedByParent` must be an independent node in the widget tree.
+// Therefore, any modifications to its child should terminate at
+// `OnlySizedByParent`. Otherwise, if its host is dirty, it implies that the
+// `OnlySizedByParent` node is also dirty, and its parent must be marked as
+// dirty. For instance, if `w2` in a Row[w1, OnlySizedByParent<w2>] is dirty,
+// the Row requires a relayout.
 impl Render for OnlySizedByParent {
   fn perform_layout(&self, clamp: BoxClamp, ctx: &mut LayoutCtx) -> Size {
     ctx
@@ -14,10 +20,6 @@ impl Render for OnlySizedByParent {
   }
 
   fn only_sized_by_parent(&self) -> bool { true }
-
-  fn paint(&self, _: &mut PaintingCtx) {
-    // nothing to paint.
-  }
 }
 
 #[cfg(test)]